+++ src/Cdn/test/New-AzCdnEndpoint.Tests.ps1
@@ -35,14 +35,16 @@ Describe 'New-AzCdnEndpoint' -Tag 'LiveOnly' {
                     HostName = "host1.hello.com"
                 };
                 $location = "westus"
+                $optimizationType = "GeneralWebDelivery"
                 Write-Host -ForegroundColor Green "Create endpointName : $($endpointName), origin.Name : $($origin.Name), origin.HostName : $($origin.HostName)"
 
-                $endpoint = New-AzCdnEndpoint -Name $endpointName -ResourceGroupName $ResourceGroupName -ProfileName $cdnProfileName -Location $location -Origin $origin
+                $endpoint = New-AzCdnEndpoint -Name $endpointName -ResourceGroupName $ResourceGroupName -ProfileName $cdnProfileName -Location $location -Origin $origin -OptimizationType $optimizationType
                 
                 $endpoint.Name | Should -Be $endpointName
                 $endpoint.Location | Should -Be $location
                 $endpoint.Origin.Name | Should -Be $origin.Name
                 $endpoint.Origin.HostName | Should -Be $origin.HostName
+                $endpoint.OptimizationType | Should -Be $optimizationType
             } Finally
             {
                 Remove-AzResourceGroup -Name $ResourceGroupName -NoWait